fix (script.js): ViewModel's public "validateUserFunction" will only attempt to validate a user function
before this change, the function had too many responsibilities; it would try to validate an acronym string corresponding to a user function, then validate for integer indexes corresponding to content items. the pivotal issue is that not every user function will need content items to execute properly. thus, this change re-implements "validateUserFunction" so it parses input for a valid user function acronym string, then executes the user function with a spread array of the rest of the input (split by a delimiter) as arguments. further validation is handled by the user function.