fix (script.js): ViewModel's public "validateUserFunction" will pass rest of user input (before first comma delimiter) to matched user function
before this change, the function would match a user function acronym string, then remove that portion (before the first delimiter) of the user input and pass the rest to the user function. this was a bug because it meant that strings with one portion (no delimiter) could not pass data to a user function.