feat (script.js) modify View's private "prepareContent" implementation
the new implementation accepts the "content" of a given state as a single value, and parses it to discover whether it contains arguments for a "parametersContainer" or "contentTable". if "content" is a string, a paragraph of text is returned. to support the new function signature of "prepareContent", the way the "render" function calls it has changed; instead of spreading the "content" argument of the state to call "prepareContent", it is directly passed