fix (script.js; Model.d.ts): flattenScheduleTreeToActivitiesArray
[script.js] update Model's private "flattenScheduleTreeToActivitiesArray" behavior
[types/Model.d.ts] update Private "flattenScheduleTreeToActivitiesArray" signature and docstring

This commit fixes the misleading docstring. "flattenScheduleTreeToActivitiesArray" did not return an array "sorted with event occurring earliest in time at the beginning", as was stated in the docstring. When applied, this commit changes that documentation and removes the second parameter; the responsibility of ordering the data is delegated to higher-level function calls.